requirements.txt
Flask==3.0.2
pymongo==4.6.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
python-dotenv==1.0.1
google-generativeai==0.4.0
google-api-python-client==2.118.0
regex==2023.12.25
beautifulsoup4==4.12.3
